refactor: rename and restructure libs/ and utils/ and add db/
