+++ src/components/ui/resizable.tsx
@@ -3,7 +3,7 @@
 import { GripVertical } from "lucide-react"
 import * as ResizablePrimitive from "react-resizable-panels"
 
-import { cn } from "@/lib/utils"
+import { cn } from "@/utils/cn"
 
 const ResizablePanelGroup = ({
   className,
